+++ test/core/elab_task.sv
@@ -6,5 +6,6 @@ module top;
     $error;
     $error("%b", 3);
     $fatal;
-    $fatal("%b", 4);
+    $fatal(0);
+    $fatal(1, "%b", 4);
 endmodule